feat: Add User() and Group() functions to FileInfo
= Added

- Add `User()` function to `Stat()`s `FileMode` struct, to return the file's
  user's name.
- Add `Group()` function to `Stat()`s `FileMode` struct to return the file's
  group name.
- Add `Error()` function to `Stat()`s `FileMode` struct to return the last error
  received when calling `BTime()`, `CTime()`, `UID()`, `GID()`,
  `User()`, or `Group()` if an additional system call is executed.
- Add `UserIDSource()` function to return if the `UID()` function or the
  `User()` function is the user's actual ID in the OS, or if the OS doesn't use
  user IDs.
- Add `PartitionType()` function.
- Add `Getuid()` and `Getgid()` functions.
- Add `IsX86CPU`, `IsArmCPU`, `IsMipsCPU`, and `IsPpcCPU` constants.

= Changed

- Change `Stat()`'s `UID()` and `GID()` to return int values, to be more
  aligned with linux/unix. ***BREAKING CHANGE***
- Speed up `Stat()` calls by deferring additional system calls to execute the
  first time `BTime()`, `CTime()`, `UID()`, `GID()`, `User()`, or `Group()` is
  called.

+++ getuid_other.go
@@ -5,16 +5,10 @@
 
 package compat
 
-// Getuid returns the User ID as a uint64. On Windows, the user's SID is
-// converted to it's POSIX equivalent, which is compatible with Cygwin and
-// Git for Windows. On Plan9, the User ID is a 64-bit hash of the user's name.
-func Getuid() (uint64, error) {
+func Getuid() (int, error) {
 	// this will intentionally not compile to alert us to a new build platform.
 }
 
-// Getgid returns the Group ID as a uint64. On Windows, the user's primary group's
-// SID is converted to its POSIX equivalent, which is compatible with Cygwin and
-// Git for Windows. On Plan9, the Getgid returns the value returned by Getuid().
-func Getgid() (uint64, error) {
+func Getgid() (int, error) {
 	// this will intentionally not compile to alert us to a new build platform.
 }
